vcs: Use a thread scoped cache invalidation context to cache repository objects.
Without this change the cache is on a process scope. If running with multiple
threads this leads to sharing the cached object between threads. This will cause
exceptions if multiple threads are trying to access the same curl object.
Even worse it allows multiple threads to operate on the same repository object
concurrently.