auth: Fix password_changed function, fixes #4043.
Never repot a changed password for default or anonymous users.

If anonymous access is disabled we don't get the default user here so we
also have to check if it is the anonymous user. In both cases (default user
and anonymous user) we can skip the password change check and return False.