svn: Set authentication realm to None in apache config template
Inside of the vcs middlewares we are using the empty string as
authentication realm for HTTP basic auth. Because we are proxy
the requestst between svn client and mod_dav_svn wen have to
use the same authentication relam in mod_dav_svn and vcs middleware.
Otherwise auth is impossible because of mismatching realms.