api: comment_pull_request, added commit_id parameter to validate status changed on particular commit.
- when using the API and creating a comment with status change it's now possible to pass in
commit_id, this will allow validation if status change of pull_request is allowed based on the given
commit_id. This solves the case when long running test sends approval of pull request which was
already updated several times. The commit_id will now validate for which state the approval was made,
and prevent accidental aproval of outdated pull requests.